update 11ty and build steps
