+++ _includes/layouts/default.html
@@ -35,10 +35,6 @@ footer, etc.).
 
   {% include "page-outdated.html" %}
 
-  {% if travel_advisory %}
-    {% include "travel-alert.html" %}
-  {% endif %}
-
   {{ content }}
   {% if words.size > site.inpagenavwordcount %}
