--- _includes/travel-alert.html
@@ -1,7 +1,0 @@
-{% capture travel_content %}
-  {% renderTemplate 'liquid' %}
-    On July 13, 2021, GSA published <a href="{% page "/july-13-travel-guidance/" %}">updated travel guidance</a> for all employees.
-  {% endrenderTemplate %}
-{% endcapture %}
-
-{% include "alert.html" level:"info" heading:"Updated Travel Guidance for GSA Employees (07/13/2021)" content:travel_content %}
