+++ package.json
@@ -13,12 +13,12 @@
     "dev": "npm run clean && npm-run-all -p dev:assets dev:serve",
     "dev:clean": "npm run clean && npm run dev",
     "dev:assets": "npm run assets:refresh && npm run assets:watch",
-    "dev:debug": "DEBUG=* npx @11ty/eleventy --serve --watch",
-    "dev:serve": "npx @11ty/eleventy --serve --watch",
+    "dev:debug": "DEBUG=* npx @11ty/eleventy --serve",
+    "dev:serve": "npx @11ty/eleventy --serve --incremental",
     "federalist": "npm run build",
     "lint": "prettier -w .",
-    "serve": "npx @11ty/eleventy --serve",
-    "start": "npx @11ty/eleventy --serve",
+    "serve": "npx @11ty/eleventy --serve --incremental",
+    "start": "npx @11ty/eleventy --serve --incremental",
     "test": "npm run test:prefixed-links && npm run test:internal-links && npm run test:html-validation",
     "test:html-validation": "html-validate _site/**/*.html",
     "test:internal-links": "node check-links.js",
@@ -27,8 +27,8 @@
   "author": "",
   "license": "CC0-1.0",
   "devDependencies": {
-    "@11ty/eleventy": "^1.0.0",
-    "@11ty/eleventy-img": "^2.0.1",
+    "@11ty/eleventy": "^2.0.1",
+    "@11ty/eleventy-img": "^3.1.0",
     "autoprefixer": "^10.4.2",
     "cheerio": "^1.0.0-rc.12",
     "chokidar-cli": "^3.0.0",
@@ -47,6 +47,7 @@
   },
   "dependencies": {
     "@uswds/uswds": "^3.2.0",
+    "glob": "^8.1.0",
     "jsdom": "^21.1.2",
     "markdown-it-attrs": "^4.1.6"
   },
