+++ pages/travel-and-leave/travel-guide-table-of-contents.md
@@ -5,7 +5,6 @@ questions:
   - tts-travel@gsa.gov
   - text: Book office hours
     url: https://sites.google.com/a/gsa.gov/tts-office-hours/
-travel_advisory: true
 redirect_from:
   - /travel-guide-table-of-contents/
 sidebar: travel
